feat/활동리스트 정렬 변경 및 리스트 컴포넌트 리팩토링
